move LoginActivity and User to correct packages
