+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity"></activity>
+        <activity android:name=".Activity.LoginActivity"></activity>
     </application>
 
 </manifest>
